Refactor and add jshint
This factors the code so that it's a little more readable. Also adds
jshint.
