+++ .editorconfig
@@ -3,3 +3,7 @@
 end_of_line = LF
 indent_style = tab
 indent_size = 4
+
+[{*.json,*.yml}]
+indent_style = space
+indent_size = 2
