+++ .jshintrc
@@ -1,0 +1,51 @@
+{
+	"globals": {
+		"steal": true,
+		"can": true,
+		"QUnit": true,
+		"System": true,
+		"SimpleDOM": true,
+		"test": true,
+		"asyncTest": true,
+		"expect": true,
+		"module": true,
+		"ok": true,
+		"equal": true,
+		"notEqual": true,
+		"deepEqual": true,
+		"propEqual": true,
+		"notDeepEqual": true,
+		"strictEqual": true,
+		"notStrictEqual": true,
+		"raises": true,
+		"start": true,
+		"stop": true,
+		"global": true,
+		"Promise": true,
+		"Symbol": true,
+		"ActiveXObject": true
+	},
+	"strict": false,
+	"curly": true,
+	"eqeqeq": true,
+	"freeze": true,
+	"indent": 2,
+	"latedef": true,
+	"noarg": true,
+	"undef": true,
+	"unused": "vars",
+	"trailing": true,
+	"maxdepth": 4,
+	"boss" : true,
+	"eqnull": true,
+	"evil": true,
+	"loopfunc": true,
+	"smarttabs": true,
+	"maxerr" : 200,
+	"jquery": true,
+	"dojo": true,
+	"mootools": true,
+	"browser": true,
+	"phantom": true,
+	"node": true
+}
