+++ lib/html-rewriter.js
@@ -6,8 +6,6 @@ var path = require("path");
 var readFile = asap(fs.readFile);
 var writeFile = asap(fs.writeFile);
 
-module.exports = HTMLRewriter;
-
 function HTMLRewriter(options){
 	options = options || {};
 	this.options = options;
@@ -46,12 +44,12 @@ HTMLRewriter.prototype = {
 
 			if(stealScript) {
 				stealScript.attr("env", "electron-production");
-				return writeFile(prodHtmlPath, $.html(), "utf8")
+				return writeFile(prodHtmlPath, $.html(), "utf8");
 			}
 		}.bind(this))
 		.catch(function(){
 			// We don't want the copy to fail just because of this
-		})
+		});
 
 	},
 
@@ -74,3 +72,5 @@ HTMLRewriter.prototype = {
 		return path.join(this.root, "production.html");
 	}
 };
+
+module.exports = HTMLRewriter;
