+++ package.json
@@ -4,7 +4,9 @@
   "description": "Build Electron applications with StealJS",
   "main": "lib/main.js",
   "scripts": {
-    "test": "mocha test/test.js",
+    "jshint": "jshint lib/. --config",
+    "mocha": "mocha test/test.js",
+    "test": "npm run jshint && npm run mocha",
     "publish": "git push origin && git push origin --tags",
     "release:pre": "npm version prerelease && npm publish",
     "release:patch": "npm version patch && npm publish",
@@ -29,11 +31,11 @@
     "cheerio": "^0.22.0",
     "electron-packager": "^8.5.0",
     "fs-extra": "^2.0.0",
-    "pdenodeify": "^0.1.0",
-    "rimraf": "^2.4.0"
+    "pdenodeify": "^0.1.0"
   },
   "devDependencies": {
     "copy": "^0.3.0",
+    "jshint": "^2.9.1",
     "mocha": "^3.2.0"
   }
 }
