+++ templates/electron-main.js
